file names ending with taxa.tsv; old util classes moved to legacy
